requirements/local.txt
-r base.txt
django-extensions>=3.2,<4.0
ipython>=8.0,<9.0